wl12xx: Add clocks overriding module parameters
It can be convenient to be able to override both the FREF and the TCXO
clock settings specified in the wl12xx_platform_data structure. This is
especially true for platforms that have more than one type of wl12xx
module.

Add two module parameters, fref & tcxo, for overriding the relevant
settings in wl12xx_platform_data.

Signed-off-by: Ido Yariv <ido@wizery.com>