mac80211/wl12xx: support scanning while idle
The current mac80211 flow goes to idle-off before starting
a scan. this in case starts the device role and ROC on it,
which complicates our state-machine pretty much.

The things gets even more complicated when considering
sched_scan, as it doesn't make mac80211 goes to idle-off
(and on issuing a normal scan while there's an ongoing
sched scan, the ROC(dev_role) will timeout).

Try simplifying it by allowing simple scan while idle.

(not sure it's upstreamable right now, but there's
an ongoing redesign of major parts of mac80211 anyway)