wl12xx: configure packet filters regardless of suspend
Enable configuring of packet filters that will be applied at
all times (not just in suspend).
When suspending, these would serve as wakeup patterns
and would wake up the host. Once activcated, any packets
which don't match would be dropped.
Therefore it's imperative to configure a unicast filter
once the mechanism is activated in order to get unicast packet
directed at the device.